Style.cpp/h rename to FillStyle.cpp/h, fix code style

+++ Lab7/composite_lib/ICanvas.h
@@ -10,7 +10,12 @@ public:
 	virtual void SetLineThickness(float thickness) = 0;
 	virtual void BeginFill(RGBAColor color) = 0;
 	virtual void EndFill() = 0;
-	virtual void MoveTo(double x, double y) = 0;
-	virtual void LineTo(double x, double y) = 0;
-	virtual void DrawEllipse(double left, double top, double width, double height) = 0;
+	virtual void MoveTo(float x, float y) = 0;
+	virtual void LineTo(float x, float y) = 0;
+	virtual void DrawEllipse(
+		float left,
+		float top,
+		float width,
+		float height
+	) = 0;
 };
